% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeApp.R
\name{makeApp}
\alias{makeApp}
\title{Make the framework off a shiny App}
\usage{
makeApp(
  dirname = NULL,
  input = NULL,
  render = NULL,
  output = NULL,
  action = NULL,
  layout = NULL,
  mode = 0,
  filename = dirname,
  reac = FALSE
)
}
\arguments{
\item{dirname}{the name of the new directory that will
contains our shiny App if mode=2, or the name of a file is mode=1.}

\item{input}{a vector of strings describing the different
types of inputs you want in the user interface. Are implemented :
'numericInput', 'sliderInput', 'checkboxInput',  'checkboxGroupInput', 'fileInput', 'downloadButton', 'radioButtons',
'dateInput', 'dateRangeInput', 'selectInput', 'textInput'.}

\item{render}{a vector of render*() the server will produce,
usually and automatically paired with output arguments.
Are implemented : 'renderImage', 'renderText', 'renderTable',
'renderDataTable', 'renderPlot'.}

\item{output}{a vector of strings describing the different
types of outputs you want in the user interface and to the
server. Are implemented : 'plotOutput', 'htmlOutput', 'imageOutput',
'tableOutput', 'dataTableOutput'.}

\item{action}{}

\item{layout}{a string among 'sidebarPanel' or 'tabsetPanel'.
Are implemented :  'sidebarLayout', 'tabsetPanel'.}

\item{mode}{0, 1 or 2, to write the code of shiny app in
\itemize{
\item 0 : the console,
\item 1 : one file, or
\item 2 : in 2 files ui.R and server.R in the dirname directory
in different files.
}}

\item{filename}{name of the file if mode=1.}

\item{reac}{boolean, TRUE or FALSE to convert on purpose every input in
reactive object, or not.}
}
\description{
\code{makeApp()} write the framework of your shiny application.
You precise the different component you need, looking at
shiny cheat sheet. Once \code{makeApp()} have produced the
framework, You'll copy and paste the recurrent codes, and adapt the
names of lists components.
}
\examples{
makeApp(output='plotOutput', mode=0, reac=TRUE)
makeApp(
  output=c('plotOutput', 'textOutput'), 
  input=c('sliderInput','numericInput'), mode=0, reac=FALSE)
}
\author{
Julien Bousquet (2023)
}
